# apps/backend/openapi.yaml
openapi: "3.0.3"
info:
  title: ArogyaAI Backend API
  version: "1.0.0"
  description: >
    Production-grade OpenAPI specification for ArogyaAI — a health-tech platform.
    Covers authentication, user management, health profiles, data ingestion,
    medical records, AI analysis, RAG-powered explanations, and alert management.
  contact:
    name: ArogyaAI Engineering
    email: engineering@arogyaai.com
  license:
    name: Private
servers:
  - url: http://localhost:8000
    description: Local Development
  - url: https://api.arogyaai.com
    description: Production

# ============================================================
# SECURITY SCHEME
# ============================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token issued at login or refresh.

  # ============================================================
  # REUSABLE SCHEMAS
  # ============================================================
  schemas:
...
    UUIDField:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    ErrorDetail:
      type: object
      properties:
        detail:
          type: string
          example: "Validation error or descriptive message"

    # --- Auth ---
...
    UserResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUIDField'
        email:
          type: string
          format: email
        full_name:
          type: string
        is_email_verified:
          type: boolean
        is_onboarding_done:
          type: boolean
        created_at:
          type: string
          format: date-time
    UpdateUserRequest:
      type: object
      properties:
        full_name:
          type: string
          example: "John Updated"
        is_onboarding_done:
          type: boolean

    # --- Health Profile ---
...
    ReportResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUIDField'
        user_id:
          $ref: '#/components/schemas/UUIDField'
        report_type:
          type: string
          enum: [BLOOD_TEST, MRI, XRAY, PRESCRIPTION, CLINICAL_NOTE, GENETIC, OTHER]
        name:
          type: string
          example: "blood_test_may.pdf"
        file_name:
          type: string
          example: "blood_test_may.pdf"
        original_filename:
          type: string
          example: "blood_test_may.pdf"
        stored_filename:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000-blood_test_may.pdf"
        file_url:
          type: string
          format: uri
          example: "https://cdn.arogyaai.com/reports/a1b2c3.pdf"
        status:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, FAILED]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        summary:
          type: array
          items:
            type: string
          example:
            - "Parsed 2 key measurements from the uploaded report."
        patient_summary:
          type: string
          example: "Parsed 2 key measurements from the uploaded report."
        summary_view:
          type: object
          additionalProperties: true
        risks:
          type: array
          items:
            type: string
        risk_level:
          type: string
          example: "Low"
        recommendations:
          type: array
          items:
            type: string
        abnormal_values:
          type: array
          items:
            type: object
            additionalProperties: true
        ocr_text:
          type: string
        summary_source:
          type: string

    # --- Risk Score ---
...
    ExplainRequest:
      type: object
      required: [report_id]
      properties:
        report_id:
          $ref: '#/components/schemas/UUIDField'
    ExplainResponse:
      type: object
      properties:
        report_id:
          $ref: '#/components/schemas/UUIDField'
        factors:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: "Systolic Blood Pressure"
              impact:
                type: string
                example: "+18%"
              direction:
                type: string
                enum: [POSITIVE, NEGATIVE, NEUTRAL]
        summary:
          type: string
          example: "Elevated diastolic pressure and elevated triglycerides are the top contributors to your risk."

    # --- Alerts ---
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
          example:
            detail: "Not authenticated"
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
          example:
            detail: "Field 'email' is required"
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
          example:
            detail: "Internal server error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
          example:
            detail: "Resource not found"

# ============================================================
# GLOBAL SECURITY (override per endpoint if public)
# ============================================================
security:
  - BearerAuth: []

# ============================================================
# PATHS / ENDPOINTS
# ============================================================
paths:
...
  /profile:
    get:
      tags: [Users]
      summary: Return the canonical normalized profile bundle for the authenticated user
      responses:
        '200':
          description: Canonical profile bundle
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: string
                  source:
                    type: string
                  error:
                    nullable: true
                  data:
                    type: object
                    properties:
                      user:
                        type: object
                      profile:
                        type: object
                      onboarding:
                        type: object
                      medical_history:
                        type: object
                      wearable:
                        type: object
                      settings:
                        type: object
                      preferences:
                        type: object
                      health_baseline:
                        type: object
                  last_updated:
                    type: string
                    nullable: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # ---- USERS ----
  /users/me:
    get:
      tags: [Users]
      summary: Get the currently authenticated user's profile
      responses:
        '200':
          description: User data returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Users]
      summary: Update the authenticated user's profile fields
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Users]
      summary: Soft-delete the currently authenticated user account
      responses:
        '200':
          description: Account deleted
          content:
            application/json:
              example:
                message: "Account successfully deleted"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # ---- HEALTH PROFILE ----
...
  /reports/upload:
    post:
      tags: [Medical Records]
      summary: Upload a medical report file (PDF, image)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, report_type]
              properties:
                file:
                  type: string
                  format: binary
                report_type:
                  type: string
                  enum: [BLOOD_TEST, MRI, XRAY, PRESCRIPTION, CLINICAL_NOTE, GENETIC, OTHER]
      responses:
        '201':
          description: Report uploaded and queued for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /reports:
    get:
      tags: [Medical Records]
      summary: List all reports for the authenticated user
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, PROCESSING, COMPLETED, FAILED]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Report list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /reports/{id}:
    get:
      tags: [Medical Records]
      summary: Retrieve a specific report by UUID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUIDField'
      responses:
        '200':
          description: Report detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Medical Records]
      summary: Delete a specific report by UUID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUIDField'
      responses:
        '200':
          description: Report deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      id:
                        $ref: '#/components/schemas/UUIDField'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /reports/{id}/access:
    get:
      tags: [Medical Records]
      summary: Resolve a temporary signed access URL for a private medical report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUIDField'
      responses:
        '200':
          description: Temporary signed access URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: string
                  error:
                    nullable: true
                  data:
                    type: object
                    properties:
                      report_id:
                        $ref: '#/components/schemas/UUIDField'
                      file_name:
                        type: string
                      url:
                        type: string
                      expires_at:
                        type: string
                      expires_in:
                        type: integer
                      legacy_public_url:
                        type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ---- ANALYSIS / PREDICTION ----
...
  /explain/report:
    post:
      tags: [RAG Explanation]
      summary: Generate a natural language explanation of a risk score via RAG pipeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExplainRequest'
      responses:
        '200':
          description: Explanation generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExplainResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...